requirements.txt
google-cloud-pubsub==2.6.1
tweepy==3.10.0